requirements.txt
asgiref==3.8.1
diff-match-patch==20230430
Django==4.2.6
django-fontawesome-5==1.0.18
django-import-export==4.1.1
djangorestframework==3.15.2
et-xmlfile==1.1.0
openpyxl==3.1.5
sqlparse==0.5.1
tablib==3.5.0
typing_extensions==4.12.2
